+++ pages/index.tsx
@@ -1,14 +1,21 @@
 // import SpinnerIllustration from "@public/vector/spinner.svg";
-import { CustomButton } from "@components/button";
+import { HomeButton } from "@components/button/home";
+import { CustomImage } from "@components/image";
 import { Layout } from "@components/layout";
 import { Section } from "@components/section";
 
-import { NextPage } from "next";
+import { domain } from "@util/config";
+import { getPreviewImage } from "@util/preview-image";
+
+import { GetStaticProps, NextPage } from "next";
 import { NextSeo } from "next-seo";
+import { PreviewImage } from "notion-types";
 
-import SpinnerIllustration from "@public/vector/spinner.svg";
+type Props = {
+    preview: PreviewImage;
+};
 
-const Home: NextPage = () => {
+const Home: NextPage<Props> = ({ preview }: Props) => {
     return (
         <Layout>
             <NextSeo
@@ -21,47 +28,75 @@ const Home: NextPage = () => {
                     },
                 ]}
             />
-            <Section className="mt-32 flex-grow">
-                <div className="w-16 absolute right-8 top-32 md:top-24">
-                    <SpinnerIllustration id="spinner" />
+            <Section className="mt-32 lg:mt-24 flex-grow mx-auto justify-center">
+                <div className="flex flex-col font-sans px-4">
+                    <h1
+                        className="text-2xl lg:text-4xl 
+                        font-light pr-12 mb-4 pb-2 w-min whitespace-nowrap border-dotted
+                        border-b-4 border-accent-alternate"
+                    >
+                        Alan John
+                    </h1>
+                    <h2 className="text-3xl md:text-7xl md:pl-4 whitespace-pre-wrap">
+                        Software Developer
+                    </h2>
                 </div>
-                <div className="py-8 my-auto mx-4 flex flex-col lg:flex-row items-center justify-evenly">
-                    <div className="flex flex-col">
-                        <div className="flex flex-col my-8">
-                            <h1 className="text-center md:text-left font-display uppercase text-7xl lg:text-8xl xl:text-9xl font-normal">
-                                Alan John
-                            </h1>
-                            <h2 className=" text-center md:text-left font-display text-3xl lg:text-4xl xl:text-5xl font-light">
-                                Software Engineer
-                            </h2>
-                        </div>
+                <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 grid-flow-row mx-auto mt-6">
+                    <div className="md:row-span-2 max-w-sm p-4 mr-4">
+                        <CustomImage
+                            alt="Alan John"
+                            src="/images/Alan.jpg"
+                            height={preview.originalHeight}
+                            width={preview.originalWidth}
+                            placeholder="blur"
+                            blurDataURL={preview.dataURIBase64}
+                        />
                     </div>
-                    <div className="flex flex-col items-center">
-                        <span className="font-display italic font-light mb-6">
-                            Select an option to proceed
-                        </span>
-                        <div className="grid grid-flow-col grid-rows-4 gap-4">
-                            <CustomButton href={"/about"} variant="retro">
-                                More about me
-                            </CustomButton>
-                            <CustomButton href={"/blog"} variant="retro">
-                                Check out Blog
-                            </CustomButton>
-                            <CustomButton
-                                href={"https://photos.alanjohn.dev"}
-                                variant="retro"
-                            >
-                                Explore Pictures
-                            </CustomButton>
-                            <CustomButton href={"/projects"} variant="retro">
-                                My Projects
-                            </CustomButton>
-                        </div>
-                    </div>
+                    <HomeButton
+                        href={"/about"}
+                        title={"about"}
+                        subtitle={"There is more to him"}
+                        description={
+                            "Considering this page does not have much about him"
+                        }
+                    />
+                    <HomeButton
+                        href={"/blog"}
+                        title={"blog"}
+                        subtitle={"He writes"}
+                        description={"He ought to be more regular with it"}
+                    />
+                    <HomeButton
+                        href={"https://photos.alanjohn.dev"}
+                        title={"Photos"}
+                        subtitle={"He clicks pictures"}
+                        description={
+                            "He is trying his best to be artsy about it"
+                        }
+                    />
+                    <HomeButton
+                        href={"/projects"}
+                        title={"projects"}
+                        subtitle={"He has projects"}
+                        description={"Surprise, surprise"}
+                    />
                 </div>
             </Section>
         </Layout>
     );
+};
+
+export const getStaticProps: GetStaticProps<Props> = async () => {
+    const preview = await getPreviewImage(domain + "/images/Alan.jpg", {
+        cacheKey: "home",
+    });
+
+    return {
+        props: {
+            preview: preview,
+        },
+        revalidate: 86400,
+    };
 };
 
 export default Home;