+++ src/components/card/blog.tsx
@@ -18,21 +18,23 @@ export const BlogCard: React.FC<Props> = ({
     return (
         <CustomLink
             href={link}
-            className="mx-1 mb-16 flex max-w-3xl flex-row border-b-2 px-2 border-secondary group"
+            className="mx-1 mb-16 flex max-w-3xl flex-row border-b-4 px-2 border-dotted border-foreground-secondary group"
         >
             <div className="my-4 px-1 flex flex-grow flex-col">
                 <h3
-                    className="font-clash text-2xl sm:text-3xl md:text-4xl lg:text-5xl 
-            group-hover:text-razzmatazz dark:group-hover:text-purple"
+                    className="font-clash text-2xl md:text-3xl lg:text-4xl 
+            group-hover:text-accent-primary dark:group-hover:text-accent-secondary"
                 >
                     {title}
                 </h3>
-                <div className="py-4 flex flex-wrap">
-                    {tags.map((tag: ITag) => (
-                        <Tag key={tag.id} {...tag} />
-                    ))}
-                </div>
-                <span className="font-sans text-jet dark:text-cultured text-lg font-light md:text-xl">
+                {tags.length > 0 && (
+                    <div className="py-3 flex flex-wrap">
+                        {tags.map((tag: ITag) => (
+                            <Tag key={tag.id} {...tag} />
+                        ))}
+                    </div>
+                )}
+                <span className="font-sans text-foreground-primary dark:text-background-primary text-lg font-light md:text-xl">
                     {description}
                 </span>
             </div>