+++ src/components/footer/index.tsx
@@ -9,7 +9,7 @@ export const Footer: React.FC = () => {
         <div className="footer">
             <CustomLink
                 href="/"
-                className="font-rubik text-xl font-semibold hidden md:block"
+                className="font-rubik text-xl font-medium hidden md:block"
             >
                 AJ
             </CustomLink>
@@ -22,11 +22,11 @@ export const Footer: React.FC = () => {
                         href={url}
                         className="p-4"
                     >
-                        <Icon className="text-2xl hover:text-orange dark:hover:text-purple" />
+                        <Icon className="text-2xl hover:text-accent-primary dark:hover:text-accent-secondary" />
                     </CustomLink>
                 ))}
             </div>
-            <div className="flex flex-col text-sm font-display text-jet dark:text-cultured">
+            <div className="flex flex-col text-sm font-sans text-foreground-secondary dark:text-background-secondary">
                 <span>Powered by Notion</span>
                 <span>Release: v{pjson.version}</span>
             </div>